requirements.txt
Package                  Version
------------------------ ----------------
absl-py                  1.4.0
adjustText               0.8
aiohttp                  3.9.1
aiosignal                1.3.1
antlr4-python3-runtime   4.8
appdirs                  1.4.4
async-timeout            4.0.3
attrs                    23.2.0
bitarray                 2.7.6
blinker                  1.6.2
cachetools               5.3.1
catboost                 1.2.2
certifi                  2023.5.7
cffi                     1.15.1
charset-normalizer       3.2.0
click                    8.1.5
cmake                    3.26.4
colorama                 0.4.6
contourpy                1.1.0
cupy-cuda117             10.6.0
cycler                   0.11.0
cysignals                1.11.4
Cython                   0.29.36
datasets                 2.1.0
decorator                4.4.2
dgl                      1.1.1+cu117
dill                     0.3.7
docker-pycreds           0.4.0
einops                   0.6.1
fairseq                  0.12.2
faiss-cpu                1.7.4
fastrlock                0.8.1
filelock                 3.12.2
fire                     0.5.0
Flask                    2.3.2
fonttools                4.41.0
frozenlist               1.4.1
fsspec                   2023.12.2
gensim                   4.3.1
gitdb                    4.0.10
GitPython                3.1.32
google-auth              2.22.0
google-auth-oauthlib     1.0.0
graphviz                 0.20.1
grpcio                   1.56.0
huggingface-hub          0.17.3
hydra-core               1.0.7
idna                     3.4
importlib-metadata       6.8.0
importlib-resources      6.0.0
itsdangerous             2.1.2
Jinja2                   3.1.2
joblib                   1.3.1
kiwisolver               1.4.4
lightgbm                 3.3.2
lightning-utilities      0.9.0
lit                      16.0.6
loguru                   0.7.0
lxml                     4.9.3
Markdown                 3.4.3
MarkupSafe               2.1.3
matplotlib               3.7.2
mpmath                   1.3.0
multidict                6.0.4
multiprocess             0.70.15
networkx                 3.1
numpy                    1.24.4
oauthlib                 3.2.2
omegaconf                2.0.6
packaging                23.1
pandas                   1.3.5
pathtools                0.1.2
Pillow                   10.0.0
pip                      23.1.2
plotly                   5.18.0
portalocker              2.7.0
protobuf                 4.23.4
psutil                   5.9.5
pyarrow                  14.0.2
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pybind11                 2.11.1
pycparser                2.21
pyparsing                3.0.9
python-dateutil          2.8.2
python-louvain           0.16
pytz                     2023.3
PyYAML                   6.0
regex                    2023.6.3
requests                 2.31.0
requests-oauthlib        1.3.1
responses                0.18.0
rsa                      4.9
sacrebleu                2.3.1
safetensors              0.4.1
scikit-learn             1.3.0
scipy                    1.11.1
sentry-sdk               1.29.2
setproctitle             1.3.2
setuptools               67.8.0
six                      1.16.0
smart-open               6.3.0
smmap                    5.0.0
spicy                    0.16.0
sympy                    1.12
tabulate                 0.9.0
tenacity                 8.2.3
tensorboard              2.13.0
tensorboard-data-server  0.7.1
termcolor                2.3.0
threadpoolctl            3.2.0
tokenizers               0.14.1
torch                    2.0.1+cu117
torch-geometric          2.3.1
torch-geometric-temporal 0.54.0
torch-scatter            2.1.1+pt20cu117
torch-sparse             0.6.17+pt20cu117
torchaudio               2.0.2
torchmetrics             1.0.1
tqdm                     4.65.0
transformers             4.35.0
triton                   2.0.0
typing_extensions        4.7.1
tzdata                   2023.3
urllib3                  1.26.16
wandb                    0.15.8
Werkzeug                 2.3.6
wget                     3.2
wheel                    0.38.4
xxhash                   3.4.1
yarl                     1.9.4
zipp                     3.16.1
